[Non braking] refactoring of some css files.
